refactor: single app profiles (moved from oc-lettings app) deleted old tables of oc-lettings and create dir for templates
