+++ lettings/urls.py
@@ -4,6 +4,6 @@ from lettings import views
 app_name = "lettings"
 
 urlpatterns = [
-    path('', views.lettings_index, name='lettings_index'),
+    path('', views.index, name='lettings_index'),
     path('<int:letting_id>/', views.letting, name='letting'),
 ]
