+++ lettings/views.py
@@ -2,10 +2,10 @@ from django.shortcuts import render
 from lettings.models import Letting
 
 
-def lettings_index(request):
+def index(request):
     lettings_list = Letting.objects.all()
     context = {'lettings_list': lettings_list}
-    return render(request, 'lettings_index.html', context)
+    return render(request, 'lettings/index.html', context)
 
 
 def letting(request, letting_id):
@@ -14,4 +14,4 @@ def letting(request, letting_id):
         'title': letting.title,
         'address': letting.address,
     }
-    return render(request, 'letting.html', context)
+    return render(request, 'lettings/letting.html', context)
