+++ oc_lettings_site/settings.py
@@ -29,6 +29,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'lettings',
+    "profiles"
 ]
 
 MIDDLEWARE = [
