--- oc_lettings_site/tests.py
@@ -1,2 +1,0 @@
-def test_dummy():
-    assert 1
