+++ oc_lettings_site/urls.py
@@ -4,9 +4,8 @@ from django.urls import path, include
 from . import views
 
 urlpatterns = [
-    path('', views.index, name='index'),
-    path('lettings/', include("lettings.urls", namespace="lettings")),
-    path('profiles/', views.profiles_index, name='profiles_index'),
-    path('profiles/<str:username>/', views.profile, name='profile'),
-    path('admin/', admin.site.urls),
+    path("", views.index, name="index"),
+    path("lettings/", include("lettings.urls", namespace="lettings")),
+    path("profiles/", include("profiles.urls", namespace="profiles")),
+    path("admin/", admin.site.urls),
 ]
