+++ oc_lettings_site/views.py
@@ -1,18 +1,5 @@
 from django.shortcuts import render
-from .models import Profile
 
 
 def index(request):
     return render(request, 'index.html')
-
-
-def profiles_index(request):
-    profiles_list = Profile.objects.all()
-    context = {'profiles_list': profiles_list}
-    return render(request, 'profiles_index.html', context)
-
-
-def profile(request, username):
-    profile = Profile.objects.get(user__username=username)
-    context = {'profile': profile}
-    return render(request, 'profile.html', context)
