+++ templates/base.html
@@ -24,7 +24,7 @@
                         <div class="container">
                             <a class="navbar-brand" href="{% url 'index'%}"><img class="img-responsive" src="{% static 'assets/img/logo.png' %}" width="70px" height="70px" alt="Logo Orange County Lettings"/></a>
                             <div>
-                                <a class="btn fw-500 ms-lg-4 btn-primary" href="{% url 'profiles_index' %}">
+                                <a class="btn fw-500 ms-lg-4 btn-primary" href="{% url 'profiles:profiles_index' %}">
                                         Profiles
                                 </a>
                                 <a class="btn fw-500 ms-lg-4 btn-primary" href="{% url 'lettings:lettings_index' %}">
