+++ templates/index.html
@@ -14,7 +14,7 @@
 
 <div class="container px-5 py-5 text-center">
     <div class="justify-content-center">
-        <a class="btn fw-500 ms-lg-4 btn-primary px-10" href="{% url 'profiles_index' %}">
+        <a class="btn fw-500 ms-lg-4 btn-primary px-10" href="{% url 'profiles:profiles_index' %}">
     		Profiles
         </a>
         <a class="btn fw-500 ms-lg-4 btn-primary px-10" href="{% url 'lettings:lettings_index' %}">
