+++ templates/lettings/index.html
@@ -1,11 +1,13 @@
 {% extends "base.html" %}
-{% block title %}Profiles{% endblock title %}
+{% block title %}Lettings{% endblock title %}
 
 {% block content %}
+
+
 <div class="container px-5 py-5 text-center">
     <div class="row justify-content-center">
         <div class="col-lg-8">
-            <h1 class="page-header-ui-title mb-3 display-6">Profiles</h1>
+            <h1 class="page-header-ui-title mb-3 display-6">Lettings</h1>
         </div>
     </div>
 </div>
@@ -14,16 +16,16 @@
     <div class="row gx-5 justify-content-center">
         <div class="col-lg-10">
             <hr class="mb-0" />
-            {% if profiles_list %}
+            {% if lettings_list %}
                 <ul class="list-group list-group-flush list-group-careers">
-                    {% for profile in profiles_list %}
+                    {% for letting in lettings_list %}
                         <li class="list-group-item">
-                            <a href="{% url 'profile' username=profile.user.username %}">{{ profile.user.username }}</a>
+                            <a href="{% url 'lettings:letting' letting_id=letting.id %}">{{ letting.title }}</a>
                         </li>
                     {% endfor %}
                 </ul>
             {% else %}
-                <p>No profiles are available.</p>
+                <p>No lettings are available.</p>
             {% endif %}
         </div>
     </div>
@@ -34,8 +36,8 @@
         <a class="btn fw-500 ms-lg-4 btn-primary px-10" href="{% url 'index' %}">
             Home
         </a>
-        <a class="btn fw-500 ms-lg-4 btn-primary px-10" href="{% url 'lettings:lettings_index' %}">
-            Lettings
+        <a class="btn fw-500 ms-lg-4 btn-primary px-10" href="{% url 'profiles:profiles_index' %}">
+            Profiles
         </a>
     </div>
 </div>
