+++ templates/lettings/letting.html
@@ -32,7 +32,7 @@
         <a class="btn fw-500 ms-lg-4 btn-primary px-10" href="{% url 'index' %}">
             Home
         </a>
-        <a class="btn fw-500 ms-lg-4 btn-primary px-10" href="{% url 'profiles_index' %}">
+        <a class="btn fw-500 ms-lg-4 btn-primary px-10" href="{% url 'profiles:profiles_index' %}">
             Profiles
         </a>
     </div>
