+++ templates/profiles/index.html
@@ -1,13 +1,11 @@
 {% extends "base.html" %}
-{% block title %}Lettings{% endblock title %}
+{% block title %}Profiles{% endblock title %}
 
 {% block content %}
-
-
 <div class="container px-5 py-5 text-center">
     <div class="row justify-content-center">
         <div class="col-lg-8">
-            <h1 class="page-header-ui-title mb-3 display-6">Lettings</h1>
+            <h1 class="page-header-ui-title mb-3 display-6">Profiles</h1>
         </div>
     </div>
 </div>
@@ -16,16 +14,16 @@
     <div class="row gx-5 justify-content-center">
         <div class="col-lg-10">
             <hr class="mb-0" />
-            {% if lettings_list %}
+            {% if profiles_list %}
                 <ul class="list-group list-group-flush list-group-careers">
-                    {% for letting in lettings_list %}
+                    {% for profile in profiles_list %}
                         <li class="list-group-item">
-                            <a href="{% url 'lettings:letting' letting_id=letting.id %}">{{ letting.title }}</a>
+                            <a href="{% url 'profiles:profile' username=profile.user.username %}">{{ profile.user.username }}</a>
                         </li>
                     {% endfor %}
                 </ul>
             {% else %}
-                <p>No lettings are available.</p>
+                <p>No profiles are available.</p>
             {% endif %}
         </div>
     </div>
@@ -36,8 +34,8 @@
         <a class="btn fw-500 ms-lg-4 btn-primary px-10" href="{% url 'index' %}">
             Home
         </a>
-        <a class="btn fw-500 ms-lg-4 btn-primary px-10" href="{% url 'profiles_index' %}">
-            Profiles
+        <a class="btn fw-500 ms-lg-4 btn-primary px-10" href="{% url 'lettings:lettings_index' %}">
+            Lettings
         </a>
     </div>
 </div>
