+++ templates/profiles/profile.html
@@ -24,7 +24,7 @@
 
 <div class="container px-5 py-5 text-center">
     <div class="justify-content-center">
-        <a class="btn fw-500 ms-lg-4 btn-primary px-10" href="{% url 'profiles_index' %}">
+        <a class="btn fw-500 ms-lg-4 btn-primary px-10" href="{% url 'profiles:profiles_index' %}">
         	<i class="ms-2" data-feather="arrow-left"></i>
             Back
         </a>
